fix: active toolbar items update on changing selection
