+++ lib/src/editor/toolbar/desktop/floating_toolbar_widget.dart
@@ -1,4 +1,5 @@
 import 'package:appflowy_editor/appflowy_editor.dart';
+import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 
 const floatingToolbarHeight = 32.0;
@@ -8,12 +9,12 @@ class FloatingToolbarWidget extends StatefulWidget {
     super.key,
     this.backgroundColor = Colors.black,
     required this.toolbarActiveColor,
-    required this.activeItems,
+    required this.items,
     required this.editorState,
     required this.layoutDirection,
   });
 
-  final List<ToolbarItem> activeItems;
+  final List<ToolbarItem> items;
   final Color backgroundColor;
   final Color toolbarActiveColor;
   final EditorState editorState;
@@ -26,7 +27,8 @@ class FloatingToolbarWidget extends StatefulWidget {
 class _FloatingToolbarWidgetState extends State<FloatingToolbarWidget> {
   @override
   Widget build(BuildContext context) {
-    if (widget.activeItems.isEmpty) {
+    var activeItems = _computeActiveItems();
+    if (activeItems.isEmpty) {
       return const SizedBox.shrink();
     }
     return Material(
@@ -37,11 +39,13 @@ class _FloatingToolbarWidgetState extends State<FloatingToolbarWidget> {
         child: SizedBox(
           height: floatingToolbarHeight,
           child: Row(
+            key: const Key('toolbar-container'),
             crossAxisAlignment: CrossAxisAlignment.start,
             mainAxisSize: MainAxisSize.min,
-            children: widget.activeItems.map((item) {
+            children: activeItems.mapIndexed((index, item) {
               final builder = item.builder;
               return Center(
+                key: Key('${item.id}-$index'),
                 child: builder!(
                   context,
                   widget.editorState,
@@ -53,5 +57,29 @@ class _FloatingToolbarWidgetState extends State<FloatingToolbarWidget> {
         ),
       ),
     );
+  }
+
+  List<ToolbarItem> _computeActiveItems() {
+    List<ToolbarItem> activeItems = widget.items
+        .where((e) => e.isActive?.call(widget.editorState) ?? false)
+        .toList();
+    if (activeItems.isEmpty) {
+      return [];
+    }
+    if (widget.layoutDirection == TextDirection.rtl) {
+      activeItems = activeItems.reversed.toList();
+    }
+    // sort by group.
+    activeItems.sort(
+      (a, b) => widget.layoutDirection == TextDirection.ltr
+          ? a.group.compareTo(b.group)
+          : b.group.compareTo(a.group),
+    );
+    // insert the divider.
+    return activeItems
+        .splitBetween((first, second) => first.group != second.group)
+        .expand((element) => [...element, placeholderItem])
+        .toList()
+      ..removeLast();
   }
 }
